Refactor the `onAddNewWebAuthnDeviceResult` method
